chore: clean up gpg keys before reinstall
Cleans up the gpg keys to allow the script to run multiple times without
issue.

+++ install/install.sh
@@ -23,9 +23,6 @@ fi
 echo "Welcome! This script will help install the BHIMA software version $BHIMA_VERSION."
 echo "=============================================================================="
 
-# show the banner image
-curl https://raw.githubusercontent.com/Third-Culture-Software/bhima-scripts/refs/heads/main/install/header.txt
-
 # Function to install dependencies
 function install_dependencies() {
   echo "Updating BHIMA OS dependencies..."
@@ -37,14 +34,19 @@ function install_dependencies() {
   echo "✓ dependencies updated"
 
   # show the banner image
+  echo "Welcome!  You are now installing"
   curl https://raw.githubusercontent.com/Third-Culture-Software/bhima-scripts/refs/heads/main/install/header.txt
+  echo "BHIMA is free and open source software (FOSS) licensed under GPLv2.  By continuing you agree to the terms of the license."
 
   # Get the LTS NodeJS from NodeSource
   curl -fsSL https://deb.nodesource.com/setup_lts.x | sudo bash -
   sudo apt-get install -y nodejs
 
   # Install the redis.io APT repository
-  rm /usr/share/keyrings/redis-archive-keyring.gpg
+  if [ -f /usr/share/keyrings/redis-archive-keyring.gpg ]; then
+    rm /usr/share/keyrings/redis-archive-keyring.gpg
+  fi
+
   curl -fsSL https://packages.redis.io/gpg | sudo gpg --dearmor -o /usr/share/keyrings/redis-archive-keyring.gpg
   sudo chmod 644 /usr/share/keyrings/redis-archive-keyring.gpg
   echo "deb [signed-by=/usr/share/keyrings/redis-archive-keyring.gpg] https://packages.redis.io/deb $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/redis.list
@@ -56,12 +58,16 @@ function install_dependencies() {
 
 # Function to install and configure MySQL
 function install_mysql() {
-
   local RELEASE_REPO="mysql-8.4-lts"
   local RELEASE_AUTH="mysql_native_password"
 
-  echo "Configuring mysql APT repository..."
+  echo "Configuring mysql APT repository... (using $RELEASE_REPO)"
+  if [ -f /usr/share/keyrings/mysql.gpg ]; then
+    rm /usr/share/keyrings/mysql.gpg
+  fi
+
   curl -fsSL https://repo.mysql.com/RPM-GPG-KEY-mysql-2023 | gpg --dearmor -o /usr/share/keyrings/mysql.gpg
+
   if [ "$(lsb_release -si)" = "Debian" ]; then
     echo "deb [signed-by=/usr/share/keyrings/mysql.gpg] http://repo.mysql.com/apt/debian $(lsb_release -sc) ${RELEASE_REPO}" >/etc/apt/sources.list.d/mysql.list
   else
